# openapi.yml
openapi: 3.0.0
info:
  title: api.josu.ai
  version: 0.2.0
...
paths:
...
  '/v0/annotations/recording/{recordingId}':
    get:
      summary: Get all annotations for given recording
      tags:
        - recording
      parameters:
        - in: path
          name: recordingId
          required: true
          schema:
            type: string
          description: Recording id
      responses:
        '200':
          description: Success
...
  '/v0/emotion/speaker/{speaker_id}':
    get:
      summary: Get emotional analysis by speaker
      tags:
        - emotions
      parameters:
        - in: path
          name: speaker_id
          required: true
          schema:
            type: string
          description: Speaker id
      responses:
        '200':
          description: Success
  '/v0/emotion/topic/{topic_id}':
    get:
      summary: Get emotional analysis by topic
      tags:
        - emotions
      parameters:
        - in: path
          name: topic_id
          required: true
          schema:
            type: string
          description: Topic id
      responses:
        '200':
          description: Success
  /v0/emotion/watson:
    post:
      summary: Get IMB Watson record
      tags:
        - emotions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  properties:
                    user_id:
                      type: string
                    provider:
                      type: string
                      enum:
                        - gmail
                        - googlecalendar
                        - instagram
                        - linkedin
                        - microsoftcalendar
                        - outlook
                        - twitter
                        - zoom
                    id:
                      type: string
                    account_id:
                      type: string
                    document_tone_ids:
                      type: array
                      nullable: true
                      items:
                        type: string
                    document_tone_names:
                      type: array
                      nullable: true
                      items:
                        type: string
                    document_tone_scores:
                      type: array
                      nullable: true
                      items:
                        type: number
                    sentences_id:
                      type: array
                      nullable: true
                      items:
                        type: number
                    sentences:
                      type: array
                      nullable: true
                      items:
                        type: string
                    sentences_score:
                      type: array
                      nullable: true
                      items:
                        type: string
                    sentences_tone_id:
                      type: array
                      nullable: true
                      items:
                        type: string
                    sentences_tone_name:
                      type: array
                      nullable: true
                      items:
                        type: string
                    speaker_id:
                      type: string
                  required:
                    - user_id
                    - provider
                    - id
                    - account_id
                    - document_tone_ids
                    - document_tone_names
                    - document_tone_scores
                    - sentences_id
                    - sentences
                    - sentences_score
                    - sentences_tone_id
                    - sentences_tone_name
                    - speaker_id
              required:
                - data
      responses:
        '201':
          description: Success
  '/v0/emotion/watson/{user_id}':
    get:
      summary: Get IBM Watson emotions by user_id
      tags:
        - emotions
      parameters:
        - in: path
          name: topic_id
          required: true
          schema:
            type: string
          description: User id
      responses:
        '200':
          description: Success
...
  '/v0/recordings/users/{username}':
    get:
      summary: Get all recordings from given speaker
      tags:
        - recordings
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: Speaker username
      responses:
        '200':
          description: Success
  '/v0/recordings/{recording_id}':
    get:
      summary: Get recording with given id
      tags:
        - recordings
      parameters:
        - in: path
          name: recording_id
          required: true
          schema:
            type: string
          description: Speaker username
      responses:
        '200':
          description: Success
    delete:
      summary: Delete recording with given id
      tags:
        - recordings
      parameters:
        - in: path
          name: recording_id
          required: true
          schema:
            type: string
          description: Speaker username
      responses:
        '200':
          description: Success
  '/v0/recordings/{recording_id}/stage':
    get:
      summary: Get all recordings by stage
      tags:
        - recordings
      parameters:
        - in: path
          name: recording_id
          required: true
          schema:
            type: string
          description: Recording id
      responses:
        '200':
          description: Success
  '/v0/recordings/{username}/processed':
    put:
      summary: Set processed field of recordings by username
      tags:
        - recordings
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: Speaker username
      responses:
        '200':
          description: Success
  '/v0/recordings/recordings/{username}':
    get:
      summary: Get all recordings queued for processing by specific user
      tags:
        - queue
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: Speaker username
      responses:
        '200':
          description: Success
...
  '/v0/users/{user_id}':
    get:
      summary: Get user by user_id
      tags:
        - users
      security:
        - Basic: []
        - Bearer: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          $ref: '#/components/responses/200_OK_User'
    put:
      summary: Update user
      tags:
        - users
      security:
        - Basic: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '202':
          $ref: '#/components/responses/202_Accepted'
...
components:
...
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUIDType'
      description: Josu user_id
...
  requestBodies:
...
    User:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/UserItem'
            required:
              - data
  responses:
...
    200_OK_User:
      description: Query successful
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                oneOf:
                  - $ref: '#/components/schemas/NullType'
                  - $ref: '#/components/schemas/UserItem'
              links:
                $ref: '#/components/schemas/Links'
            required:
              - data
              - links
...
    202_Accepted:
      description: Command successful
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                $ref: '#/components/schemas/NullType'
              links:
                $ref: '#/components/schemas/Links'
            required:
              - data
              - links
...
  schemas:
...
    Links:
      type: object
      properties:
        self:
          $ref: '#/components/schemas/URLType'
        next:
          $ref: '#/components/schemas/URLType'
        resource:
          $ref: '#/components/schemas/URLType'
      required:
        - self
...
    AccountItem:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/StringIdType'
        autojoin_zoom:
          type: boolean
        email:
          $ref: '#/components/schemas/EmailType'
        first_name:
          $ref: '#/components/schemas/FirstNameType'
        last_name:
          $ref: '#/components/schemas/LastNameType'
        name:
          $ref: '#/components/schemas/NameType'
        provider:
          $ref: '#/components/schemas/CalendarProviderType'
        refresh_token:
          $ref: '#/components/schemas/RefreshTokenType'
      required:
        - id
        - autojoin_zoom
        - email
        - first_name
        - last_name
        - name
        - provider
        - refresh_token
    CalendarItem:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/StringIdType'
        account:
          $ref: '#/components/schemas/StringIdType'
        channel:
          $ref: '#/components/schemas/NullableStringIdType'
        expires:
          $ref: '#/components/schemas/NullableDateTimeType'
        provider:
          $ref: '#/components/schemas/CalendarProviderType'
        refresh_token:
          $ref: '#/components/schemas/RefreshTokenType'
      required:
        - id
        - account
        - channel
        - expires
        - provider
        - refresh_token
    UserItem:
      type: object
      properties:
        user_id:
          $ref: '#/components/schemas/UUIDType'
        __accounts_google:
          $ref: '#/components/schemas/UserAccountsItem'
        __accounts_microsoft:
          $ref: '#/components/schemas/UserAccountsItem'
        __calendars_google:
          $ref: '#/components/schemas/UserCalendarsItem'
        __calendars_microsoft:
          $ref: '#/components/schemas/UserCalendarsItem'
        autojoin_zoom:
          type: boolean
        devices:
          type: array
          minItems: 0
          maxItems: 16
          nullable: true
          items:
            $ref: '#/components/schemas/UUIDType'
        email:
          $ref: '#/components/schemas/EmailType'
        first_name:
          $ref: '#/components/schemas/FirstNameType'
        last_name:
          $ref: '#/components/schemas/LastNameType'
        password:
          $ref: '#/components/schemas/NullablePasswordHashType'
        profile_id:
          $ref: '#/components/schemas/UUIDType'
        username:
          $ref: '#/components/schemas/NullableUsernameType'
        userpic_url:
          $ref: '#/components/schemas/NullableURLType'
      required:
        - user_id
        - __accounts_google
        - __accounts_microsoft
        - __calendars_google
        - __calendars_microsoft
        - autojoin_zoom
        - devices
        - email
        - first_name
        - last_name
        - password
        - profile_id
        - username
        - userpic_url
    UserAccountsItem:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/AccountItem'
      example:
        '115244245744751925401':
          autojoin_zoom: false
          email: vf.velt@gmail.com
          first_name: Oleksii
          last_name: Dorozhkin
          name: Oleksii Dorozhkin
          provider: google
          refresh_token: 1//0f3_7fcGe6aD0CgYIARAAGA8SNwF
    UserCalendarsItem:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CalendarItem'
      example:
        vf.velt@gmail.com:
          account: '115244245744751925401'
          channel: f86bc02d-051f-4efa-a372-80dc0b6d4235
          expires: '2020-08-09T21:00:00.000Z'
          provider: google
          refresh_token: 1//0f3_7fcGe6aD0CgYIARAAGA8SNwF
...
    CalendarProviderType:
      type: string
      enum:
        - google
        - microsoft
        - zoom
      example: google
...
    EmailType:
      type: string
      format: email
      example: vf.velt@gmail.com
...
    FirstNameType:
      type: string
      minLength: 1
      maxLength: 32
      example: Oleksii
    LastNameType:
      type: string
      minLength: 1
      maxLength: 32
      example: Dorozhkin
...
    NameType:
      type: string
      minLength: 0
      maxLength: 128
      example: Oleksii Dorozhkin
...
    NullableDateTimeType:
      type: string
      format: date-time
      nullable: true
      example: '2020-08-09T21:00:00.000Z'
...
    NullablePasswordHashType:
      type: string
      minLength: 48
      maxLength: 64
      nullable: true
      example: $2b$12$1TXi0Ro6wdmvIvtvd4g3o.MKj2SdikTBiiRaadTA9zoU0h5h5wBhm
...
    NullableStringIdType:
      type: string
      minLength: 4
      maxLength: 512
      nullable: true
      example: '103893822843346636622'
    NullableUsernameType:
      type: string
      minLength: 4
      maxLength: 32
      nullable: true
      example: my user name
    NullableURLType:
      type: string
      format: uri
      nullable: true
      example: 'https://lh3.googleusercontent.com/a-/AOh14Gjwet8KDrdXwsdyux'
...
    NullType:
      type: boolean
      nullable: true
      example: 'null'
...
    RefreshTokenType:
      type: string
      minLength: 64
      maxLength: 2048
      example: 1//0f3_7fcGe6aD0CgYIARAAGA8SNwF-L9IrVRnT3cB3kEea
    StringIdType:
      type: string
      minLength: 4
      maxLength: 512
      example: '103893822843346636622'
...
    URLType:
      type: string
      format: uri
      example: 'https://lh3.googleusercontent.com/a-/AOh14Gjwet8KDrdXwsdyux'
    UUIDType:
      type: string
      format: uuid
      example: f86bc02d-051f-4efa-a372-80dc0b6d4235
...
  securitySchemes:
    Basic:
      type: http
      scheme: basic
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    UserIdCookie:
      type: apiKey
      in: cookie
      name: userId
    UserIdSecCookie:
      type: apiKey
      in: cookie
      name: userId.sec
...